mid-migration to stop using constants globally
